Rename ssh keys view and fix delete action value
