+++ nest-bot/src/actions/delete_ssh_key.ts
@@ -4,7 +4,7 @@ import {
   get_authorized_keys,
   set_authorized_keys,
 } from "../os/os_functions.js";
-import ssh_keys_input from "../blocks/ssh_keys_input.js";
+import ssh_keys_input from "../blocks/ssh_keys_view.js";
 
 export function delete_ssh_key(app: Slack.App) {
   app.action("delete_ssh_key", async ({ ack, body, client }) => {
